Week07_Docker_Deployment/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
python-dotenv
pydantic-settings
httpx
pytest
pytest-asyncio
